refactor: cardApi 기능을 의존성 주입을 통해 HtttpCardRepository로 이관
