+++ my-app/src/repository/CardRepositoryShell.ts
@@ -1,0 +1,7 @@
+import type { Card, SendingData } from "./HttpCardRepository";
+
+export interface CardRepositoryShell {
+  getCards(): Promise<Card[]>;
+  postCard(sendingData: SendingData): Promise<{ id: string }>;
+  deleteCard(id: string): Promise<void>;
+}
