+++ my-app/src/repository/RepositoryChanger.ts
@@ -1,0 +1,3 @@
+import { HttpCardRepository } from "./HttpCardRepository";
+
+export const cardRepository = new HttpCardRepository();
